+++ uui/components/pickers/DataPickerBody.tsx
@@ -120,6 +120,8 @@ export function DataPickerBody<TItem, TId>({ highlightSearchMatches = true, ...p
 
     const renderedDataRows = useMemo(() => props.rows.map((row) => renderRow(row, props.value)), [props.rows, props.value]);
 
+    // console.log('props.rows', props.rows);
+
     return (
         <>
             {showSearch && (