+++ changelog.md
@@ -2,13 +2,14 @@
 **What's New**
 * [RangeDatePicker]: improved a11y, updated so that when a date is typed in the input fields, the calendar body immediately reflects and selects the new value, providing instant feedback and better usability.
 * [DatePicker]: improved a11y.
+* [@epam/uui-test-utils]: added global mock for "getBoundingClientRect" to "setupJsDom"
 
 **What's Fixed**
 * [RangeDatePicker]: fixed switching of the calendar panel depending on which field the action was on.
 * [PickerInput]: fixed returning to the previous focused element via shift+tab press
 * [Rating]: fixed wrong rating calculations in 'block' container and 'sticky' value on hover
 * [Modals][Eectric]: increased border radius to 12 px
-
+* [PickerModal]: disabled "Show only selected" when search is active
 
 # 6.1.0 - 08.05.2025